cleanup(core): move all models to a single file types.ts
