--- packages/core/src/lib/models/camera.ts
@@ -1,9 +1,0 @@
-import * as THREE from 'three';
-import type { NgtCommonParameters, Overwrite } from './three';
-
-export type NgtCamera = THREE.OrthographicCamera | THREE.PerspectiveCamera;
-
-export type NgtCameraOptions =
-  | NgtCamera
-  | Overwrite<Partial<THREE.PerspectiveCamera>, NgtCommonParameters>
-  | Overwrite<Partial<THREE.OrthographicCamera>, NgtCommonParameters>;
