--- packages/core/src/lib/models/dpr.ts
@@ -1,1 +1,0 @@
-export type NgtDpr = number | [min: number, max: number];
