--- packages/core/src/lib/models/index.ts
@@ -1,22 +1,0 @@
-export * from './states/state';
-export * from './states/events-state';
-export * from './states/animation-frame-state';
-export * from './states/performance-state';
-
-export * from './events';
-export * from './three';
-export * from './render';
-export * from './size';
-export * from './viewport';
-export * from './intersection';
-export * from './instance';
-export * from './raycaster';
-export * from './scene';
-export * from './camera';
-export * from './common';
-export * from './loader';
-export * from './object-map';
-export * from './object';
-export * from './dpr';
-
-export * from './options/gl-options';
