--- packages/core/src/lib/models/intersection.ts
@@ -1,25 +1,0 @@
-import * as THREE from 'three';
-
-export interface NgtIntersection extends THREE.Intersection {
-  eventObject: THREE.Object3D;
-}
-
-export interface NgtIntersectionEvent<TSourceEvent> extends NgtIntersection {
-  intersections: NgtIntersection[];
-  stopped: boolean;
-  unprojectedPoint: THREE.Vector3;
-  ray: THREE.Ray;
-  camera: THREE.Camera;
-  stopPropagation: () => void;
-  /**
-   * @deprecated use {@link nativeEvent}
-   */
-  sourceEvent: TSourceEvent;
-  nativeEvent: TSourceEvent;
-  delta: number;
-}
-
-export interface NgtPointerCaptureTarget {
-  intersection: NgtIntersection;
-  target: Element;
-}
