--- packages/core/src/lib/models/object-map.ts
@@ -1,6 +1,0 @@
-import * as THREE from 'three';
-
-export interface NgtObjectMap {
-  nodes: { [name: string]: THREE.Object3D };
-  materials: { [name: string]: THREE.Material };
-}
