--- packages/core/src/lib/models/options/gl-options.ts
@@ -1,7 +1,0 @@
-import * as THREE from 'three';
-import { Properties } from '../common';
-
-export type NgtGLOptions =
-  | THREE.Renderer
-  | ((canvas: HTMLCanvasElement) => THREE.Renderer)
-  | Partial<Properties<THREE.WebGLRenderer> | THREE.WebGLRendererParameters>;
