--- packages/core/src/lib/models/scene.ts
@@ -1,9 +1,0 @@
-import * as THREE from 'three';
-import type { NgtCommonParameters, Overwrite } from './three';
-
-export type NgtSceneOptions = Overwrite<
-  Partial<
-    Omit<THREE.Scene, 'isScene' | 'onBeforeRender' | 'onAfterRender' | 'type'>
-  >,
-  NgtCommonParameters
->;
