--- packages/core/src/lib/models/size.ts
@@ -1,4 +1,0 @@
-export interface NgtSize {
-  width: number;
-  height: number;
-}
