--- packages/core/src/lib/models/states/performance-state.ts
@@ -1,6 +1,0 @@
-export interface NgtPerformance {
-  current?: number;
-  min?: number;
-  max?: number;
-  debounce?: number;
-}
