+++ packages/core/src/lib/three/attribute.ts
@@ -8,7 +8,7 @@ import {
   Optional,
 } from '@angular/core';
 import * as THREE from 'three';
-import type { AnyConstructor } from '../models';
+import type { AnyConstructor } from '../types';
 import { NgtGeometry } from './geometry';
 
 @Directive()
