+++ packages/core/src/lib/three/geometry.ts
@@ -9,8 +9,8 @@ import {
 } from '@angular/core';
 import { requestAnimationFrame } from '@rx-angular/cdk/zone-less';
 import * as THREE from 'three';
-import type { AnyConstructor, AnyFunction, UnknownRecord } from '../models';
 import { NGT_OBJECT_3D } from '../providers/object3d';
+import type { AnyConstructor, AnyFunction, UnknownRecord } from '../types';
 
 @Directive()
 export abstract class NgtGeometry<
